Fixed horrible bug disabling startup
I had let code in @PostConstruct of OakRepository call other methods on
other SpringBeans, which may or may not be yet constructed. So the
precise ordering of SpringBean construction had crept in and was a
timebomb curcular dependency waiting to go off. It went off. But it's
fixed. Just remember the lesson NEVER leat @PostConstruc code in one
bean reference other beans, because they may NOT be constructed yet, in
which case the wirings will be null.